+++ frontend/Assets/Scripts/Battle/Weapons/Firearms/HeatPistolWeapon.cs
@@ -10,7 +10,7 @@ namespace Battle.Weapons.Firearms {
         public static WeaponDescriptor Descriptor {
             get {
                 return new WeaponDescriptor(
-                    (int) WeaponId.HeatPistol,
+                    WeaponId.HeatPistol,
                     The<WeaponIcons>.Get().HeatPistol
                 );
             }
