+++ frontend/Assets/Scripts/Battle/Weapons/OtherUtils/SkipTurnWeapon.cs
@@ -10,7 +10,7 @@ namespace Battle.Weapons.OtherUtils {
         public static WeaponDescriptor Descriptor {
             get {
                 return new WeaponDescriptor(
-                    (int) WeaponId.SkipTurn,
+                    WeaponId.SkipTurn,
                     The<WeaponIcons>.Get().SkipTurn
                 );
             }
